@@ -1,17 +1,47 @@
-source 'http://rubygems.org'
+source 'https://rubygems.org'
 
-gem 'rails', '3.1.3'
-gem 'mysql2'
+gem 'rails', '3.2.3'
 
-# asset pipeline
+# Bundle edge Rails instead:
+# gem 'rails', :git => 'git://github.com/rails/rails.git'
+
+gem 'sqlite3'
+
+
+# Gems used only for assets and not required
+# in production environments by default.
 group :assets do
-  gem 'sass-rails', '~> 3.1.5.rc.2'
-  gem 'coffee-rails', '~> 3.1.1'
+  gem 'sass-rails',   '~> 3.2.3'
+  gem 'coffee-rails', '~> 3.2.1'
+
+  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
+  # gem 'therubyracer', :platform => :ruby
+
   gem 'uglifier', '>= 1.0.3'
+
+  gem 'bootstrap-sass'
+  gem 'bourbon'
+
 end
 
 gem 'jquery-rails'
 
-# gems for omniauth and devise
-gem 'devise', :git => 'git://github.com/plataformatec/devise.git'
+
+gem 'devise'
 gem 'omniauth'
+
+
+# To use ActiveModel has_secure_password
+# gem 'bcrypt-ruby', '~> 3.0.0'
+
+# To use Jbuilder templates for JSON
+# gem 'jbuilder'
+
+# Use unicorn as the app server
+# gem 'unicorn'
+
+# Deploy with Capistrano
+# gem 'capistrano'
+
+# To use debugger
+# gem 'ruby-debug19', :require => 'ruby-debug'
