@@ -1,8 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-OauthProviderDemo::Application.config.session_store :cookie_store, :key => '_oauth-provider-demo_session'
+Ssop::Application.config.session_store :cookie_store, key: '_ssop_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
-# (create the session table with "rake db:sessions:create")
-# OauthProviderDemo::Application.config.session_store :active_record_store
+# (create the session table with "rails generate session_migration")
+# Ssop::Application.config.session_store :active_record_store
