@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-OauthProviderDemo::Application.config.secret_token = '16a18b5f4160afbfa2a0ec7b78a4903e0a63cfb819a641eaea6667be269eb13de6f8ae1073f897b735e21e488ede314d08ad856545cb8f16edd5c617cf5a00d3'
+Ssop::Application.config.secret_token = 'cee21355ad6ef926960b9c623a817fa061cffe29eab1e2eea9b3692a9b89e1caf7316fb591fab1d127470eccea60408125e4a6e5d25f24f5fcdab953aa3fa738'
