@@ -1,19 +1,78 @@
-OauthProviderDemo::Application.routes.draw do
+Ssop::Application.routes.draw do
   devise_for :users, :controllers => { :registrations => 'registrations',
                                        :sessions => 'sessions'}
   # omniauth client stuff
   match '/auth/:provider/callback', :to => 'authentications#create'
-  match '/auth/failure', :to => 'authentications#failure'
+  match '/auth/failure',            :to => 'authentications#failure'
 
   # Provider stuff
-  match '/auth/josh_id/authorize' => 'auth#authorize'
-  match '/auth/josh_id/access_token' => 'auth#access_token'
-  match '/auth/josh_id/user' => 'auth#user'
-  match '/oauth/token' => 'auth#access_token'
+  match '/auth/ssop4r/authorize'    => 'auth#authorize'
+  match '/auth/ssop4r/access_token' => 'auth#access_token'
+  match '/auth/ssop4r/user'         => 'auth#user'
+  match '/auth/ssop4r/token'        => 'auth#access_token'
 
   # Account linking
   match 'authentications/:user_id/link' => 'authentications#link', :as => :link_accounts
-  match 'authentications/:user_id/add' => 'authentications#add', :as => :add_account
+  match 'authentications/:user_id/add'  => 'authentications#add', :as => :add_account
 
   root :to => 'auth#welcome'
+
+  # root :to => "home#index"
+
+  # The priority is based upon order of creation:
+  # first created -> highest priority.
+
+  # Sample of regular route:
+  #   match 'products/:id' => 'catalog#view'
+  # Keep in mind you can assign values other than :controller and :action
+
+  # Sample of named route:
+  #   match 'products/:id/purchase' => 'catalog#purchase', :as => :purchase
+  # This route can be invoked with purchase_url(:id => product.id)
+
+  # Sample resource route (maps HTTP verbs to controller actions automatically):
+  #   resources :products
+
+  # Sample resource route with options:
+  #   resources :products do
+  #     member do
+  #       get 'short'
+  #       post 'toggle'
+  #     end
+  #
+  #     collection do
+  #       get 'sold'
+  #     end
+  #   end
+
+  # Sample resource route with sub-resources:
+  #   resources :products do
+  #     resources :comments, :sales
+  #     resource :seller
+  #   end
+
+  # Sample resource route with more complex sub-resources
+  #   resources :products do
+  #     resources :comments
+  #     resources :sales do
+  #       get 'recent', :on => :collection
+  #     end
+  #   end
+
+  # Sample resource route within a namespace:
+  #   namespace :admin do
+  #     # Directs /admin/products/* to Admin::ProductsController
+  #     # (app/controllers/admin/products_controller.rb)
+  #     resources :products
+  #   end
+
+  # You can have the root of your site routed with "root"
+  # just remember to delete public/index.html.
+  # root :to => 'welcome#index'
+
+  # See how all your routes lay out with "rake routes"
+
+  # This is a legacy wild controller route that's not recommended for RESTful applications.
+  # Note: This route will make all actions in every controller accessible via GET requests.
+  # match ':controller(/:action(/:id))(.:format)'
 end
